feat(frontend): move roulette wheel to homepage
